Update vrc-quick-launcher.md
I always thought it was an internal tool because VRChat does not mention ANYWHERE in their docs where you can get it, so I added an explanation that this tool is part of the Creator Companion and can be found inside there. When you google "VRC Quick Launcher", you can't find any info about where to find this secret tool, so that is definitely needed.